requirements.txt
Django>=4.0
markdown2
djangorestframework
django-cors-headers